requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
mysqlclient==2.0.1
django-mysql==3.8.1
Pillow ==8.1.2
cryptography==3.4.7
django-allauth == 0.44.0
dj-rest-auth==2.1.4
djangorestframework==3.12.4
drf-yasg==1.20.0
django-filter==2.4.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0



